FIXES - vip armor wrong color (or wrong color in general doesnt match teams) - arcane shield absorption hearts dont fade - arcane shield absorption hearts dont disappear at all sometimes - orbs dont despawn (dubious) - respawns face wrong direction (on warsong) - heal chain does not crit(?)  (NOT A BUG?) - can move rain while in spectator - shows wounding/cripple as a positive (should be red) - multiple links (souldbinds) dont give you multiple healings

+++ src/main/java/com/ebicep/warlords/classes/abilties/TimeWarp.java
@@ -107,9 +107,9 @@ public class TimeWarp extends AbstractAbility {
         private WarlordsPlayer warlordsPlayer;
         private Location location;
         private Vector facing;
-        private int time;
+        private float time;
 
-        public TimeWarpPlayer(WarlordsPlayer warlordsPlayer, Location location, Vector facing, int time) {
+        public TimeWarpPlayer(WarlordsPlayer warlordsPlayer, Location location, Vector facing, float time) {
             this.warlordsPlayer = warlordsPlayer;
             this.location = location;
             this.facing = facing;
@@ -140,11 +140,11 @@ public class TimeWarp extends AbstractAbility {
             this.facing = facing;
         }
 
-        public int getTime() {
+        public float getTime() {
             return time;
         }
 
-        public void setTime(int time) {
+        public void setTime(float time) {
             this.time = time;
         }
     }
